+++ src/main/java/org/flamierawieo/anirandom/orm/dao/PlanToWatchDao.java
@@ -1,0 +1,7 @@
+package org.flamierawieo.anirandom.orm.dao;
+
+public class PlanToWatchDao {
+
+    // TODO: todo
+
+}
