+++ src/main/java/org/flamierawieo/anirandom/orm/mapping/User.java
@@ -1,4 +1,4 @@
-package org.flamierawieo.anirandom.orm;
+package org.flamierawieo.anirandom.orm.mapping;
 
 import org.bson.types.ObjectId;
 import org.mongodb.morphia.annotations.Embedded;
