+++ src/main/resources/static/css/plan_to_watch.css
@@ -7,6 +7,7 @@ main .plan_to_watch {
 }
 
 main section {
+    position: relative;
     width: 120vmin;
     height: 16vmin;
     background: #FFFFFF;
@@ -40,6 +41,7 @@ main section .info {
 }
 
 main section .info .drop-down-more .more {
+    top: 1vh;
     position: absolute;
     left: 115vmin;
     background: url("../images/more.svg");
@@ -62,9 +64,9 @@ main section .info .drop-down-more .bubble {
 }
 
 main section .info .drop-down-more.show .bubble {
-    right: 4.5vmin;
+    right: 5.5vmin;
     position: absolute;
-    margin-top: 5vmin;
+    margin-top: 2vmin;
     height: 9vmin;
     width: 25vmin;
     background: white;
@@ -108,7 +110,7 @@ main section .info .stars {
     margin-left: 1.6vmin;
     background-repeat: repeat-x;
     background-image: url("../images/star.svg");
-    width: 68.3vmin;
+    width: 70.3vmin;
     height: 7vmin;
     left: 13.5vmin;
     transition: 0.2s ease;
@@ -122,7 +124,7 @@ main section .info .stars .stars-fill {
 
 main section .info .rating {
     margin-top: -5.2vmin;
-    margin-left: 71vmin;
+    margin-left: 73vmin;
     font-family: "Lato-Regular";
     font-size: 3.9vmin;
     color: #5B5B5B;
